+++ MiniChallenge2/MiniChallenge2/Views/OlympicsHistory/Components/OlympicHistoryCardDivider.swift
@@ -1,0 +1,17 @@
+//
+//  OlympicHistoryCardDivider.swift
+//  MiniChallenge2
+//
+//  Created by Rafa (Ruffles) on 12/09/23.
+//
+
+import Foundation
+import SwiftUI
+
+struct OlympicHistoryCardDivider : View {
+    var body: some View {
+        VStack{}
+            .frame(width: 84, height: 2)
+            .background(Color("Black"))
+    }
+}
